orders/src/events/listeners/__test__/expiration-complete-listener.test.ts
import mongoose from 'mongoose';
import { Message } from 'node-nats-streaming';
import { OrderStatus, ExpirationCompleteEvent } from '@myappsack/common-features';
import { ExpirationCompleteListener } from '../expiration-complete-listener';
import { natsWrapper } from '../../../nats-wrapper';
import { Order } from '../../../models/order';
import { Ticket } from '../../../models/ticket';

const setup = async () => {
  const listener = new ExpirationCompleteListener(natsWrapper.client);
  const ticketid = new mongoose.Types.ObjectId().toHexString();  
  const ticket = Ticket.build({
    id: ticketid,
    title: 'concert',
    price: 20,
  });
  await ticket.save();
  const order = Order.build({
    status: OrderStatus.Created,
    userId: 'alskdfj',
    expiresAt: new Date(),
    ticket,
  });
  await order.save();


  const data: ExpirationCompleteEvent['data'] = {
    orderId: order.id,
  };

  // @ts-ignore
  const msg: Message = {
    ack: jest.fn(),
  };

  return { listener, order, ticket, data, msg };
};

it('updates the order status to cancelled', async () => {
  const { listener, order, data, msg } = await setup();

  console.log('orderid 1:', order.id);
  console.log('data: 1', data);

  await listener.onMessage(data, msg);

  const updatedOrder = await Order.findById(order.id);
  expect(updatedOrder!.status).toEqual(OrderStatus.Cancelled);
});



it('ack the message', async () => {
  const { listener, data, msg } = await setup();

  console.log('data: 2', data);

  await listener.onMessage(data, msg);

  expect(msg.ack).toHaveBeenCalled();
});

it('emit an OrderCancelled event', async () => {
    const { listener, order, data, msg } = await setup();
    await listener.onMessage(data, msg);
    expect(natsWrapper.client.publish).toHaveBeenCalled();
  
    const eventData = JSON.parse(
      (natsWrapper.client.publish as jest.Mock).mock.calls[0][1]
    );
  
    console.log('eventid: 3', eventData.id);
    console.log('orderid: 3', order.id);
    console.log('data: 3', data);
  
    expect(eventData.id).toEqual(order.id);
  });
